+++ internal/cli/atlas/maintenance/update.go
@@ -56,8 +56,8 @@ func (opts *UpdateOpts) Run() error {
 
 func (opts *UpdateOpts) newMaintenanceWindow() *atlasv2.GroupMaintenanceWindow {
 	return &atlasv2.GroupMaintenanceWindow{
-		DayOfWeek: int32(opts.dayOfWeek),
-		HourOfDay: int32(opts.hourOfDay),
+		DayOfWeek: opts.dayOfWeek,
+		HourOfDay: opts.hourOfDay,
 		StartASAP: &opts.startASAP,
 	}
 }
