+++ internal/kubernetes/operator/deployment/deployment_test.go
@@ -136,7 +136,7 @@ func TestBuildAtlasAdvancedDeployment(t *testing.T) {
 			SampleSizeBIConnector:            pointer.Get[int32](10),
 			SampleRefreshIntervalBIConnector: pointer.Get[int32](10),
 		}
-		processArgs.OplogSizeMB.Set(pointer.Get(int32(10)))
+		processArgs.OplogSizeMB.Set(pointer.Get(10))
 		processArgs.OplogMinRetentionHours.Set(pointer.Get(float64(10.1)))
 		backupSchedule := &atlasv2.DiskBackupSnapshotSchedule{
 			ClusterId:             pointer.Get("testClusterID"),
